Move performance scenario in its own file
